move mempool to its new directory
