+++ common/log/logger.cpp
@@ -254,3 +254,4 @@ loglevel logger::adjust_level() {
 }
 
 
+
