+++ common/mempool/mempool.cpp
@@ -16,7 +16,7 @@
     License along with this library.
 */
 
-#include <mempool.hpp>
+#include <mempool/mempool.hpp>
 
 #include <unordered_map>
 #include "buffer.hpp"
